+++ language/dutch_utf-8.php
@@ -387,10 +387,14 @@ $LANG_confignames['evlist'] = array(
     'cal_tmpl'              => 'Use HTML or JSON templates',
     'enable_rsvp'           => 'Enable Registration/Ticketing?',
     'rsvp_print'            => 'Enable Ticket Printing?',
+    'meetup_key'            => 'Meetup API Key',
+    'meetup_gid'            => 'Meetup Group ID(s)',
+    'meetup_cache_minutes'  => 'Cache Minutes',
 );
 $LANG_configsubgroups['evlist'] = array(
     'sg_main'               => 'Hoofd Instellingen',
     'sg_rsvp'               => 'RSVP/Ticketing',
+    'sg_integ'              => 'Integrations',
 );
 $LANG_fs['evlist'] = array(
     'ev_access'             => 'Toegangscontrole',
@@ -398,6 +402,7 @@ $LANG_fs['evlist'] = array(
     'ev_centerblock'        => 'Centerblok Instellingen',
     'ev_permissions'        => 'Standaard Rechten',
     'ev_rsvp'               => 'Registration and Ticketing',
+    'ev_integ_meetup'       => 'Meetup.Com',
 );
 $LANG_configselects['evlist'] = array(
     0 => array('Ja' => 1, 'Nee' => 0),