+++ language/german_formal.php
@@ -386,10 +386,14 @@ $LANG_confignames['evlist'] = array(
     'cal_tmpl'              => 'Use HTML or JSON templates',
     'enable_rsvp'           => 'Enable Registration/Ticketing?',
     'rsvp_print'            => 'Enable Ticket Printing?',
+    'meetup_key'            => 'Meetup API Key',
+    'meetup_gid'            => 'Meetup Group ID(s)',
+    'meetup_cache_minutes'  => 'Cache Minutes',
 );
 $LANG_configsubgroups['evlist'] = array(
     'sg_main'               => 'Haupteinstellungen',
     'sg_rsvp'               => 'RSVP/Ticketing',
+    'sg_integ'              => 'Integrations',
 );
 $LANG_fs['evlist'] = array(
     'ev_access'             => 'Zugriffskontrolle',
@@ -397,6 +401,7 @@ $LANG_fs['evlist'] = array(
     'ev_centerblock'        => 'Centerblock-Einstellungen',
     'ev_permissions'        => 'Standardberechtigungen',
     'ev_rsvp'               => 'Registration and Ticketing',
+    'ev_integ_meetup'       => 'Meetup.Com',
 );
 $LANG_configselects['evlist'] = array(
     0 => array('Ja' => 1, 'Nein' => 0),